Large update to Dendrobands
Establishing the potential format of the Dendroband repo files.

Added in this commit are:
Biannual data, with forms of
1. Current year data
2. Map of tree locations
3. Field form
4. Species list
5. Old data sheet for reference

Intraannual data, with forms of
1. Current year data
2. Map of tree locations
3. Field form
4. Species list
5. Old data sheet for reference

Master data file

Metadata, including 4 metadata files describing the above data files.

I will be updating the README for the repo as well as creating some issues/questions to address points of clarification.

@teixeirak @gonzalezeb these files are ready to be reviewed before we start filling them out with data. I'll comment on this commit when I've finished updating the README and putting in my issues.