Convert 2017 data to new format
This batch upload consists of changed files in the course of making the 2017 data be in the new format. Scripts were also reorganized and edited, and separate data entry forms were made for conversion in R.

Dendro_trees was updated with more accurate mortality dates.

Nothing was changed in the archive data, but as excel froze and crashed as was working with it, it saved a "recovered copy" before I could get rid of it. Hence, it has a new save date.