don't run the "testthat" so GitHub actions don't fail (+ trigger actions)
This is so that if GitHub action don't fail if everything ran smoothly. If there are issues to address in the data, that is given by the dashboard and the report. This test that makes it look that actions failed, but there could be a lot of reasons for that (unrelated to error in the data).